Total concentrate feed function added
Added a function (from line 261) that computes the concentrate needs from each of the three farm types. I tried to keep this in the same structure as previous calculations, so I added a function in the cattle and sheep LCA tools that quickly outputs the concentrates per animal and iterates through that.